introduce MapResponseParser and StreamingResponseParser
