+++ apollo-integration/src/test/kotlin/com/apollographql/apollo/performance/NumberParsingTest.kt
@@ -1,20 +1,11 @@
 package com.apollographql.apollo.performance
 
 import com.apollographql.apollo.api.parse
-import com.apollographql.apollo.ApolloClient
-import com.apollographql.apollo.Utils.immediateExecutor
-import com.apollographql.apollo.Utils.immediateExecutorService
-import com.apollographql.apollo.api.internal.SimpleOperationResponseParser
 import com.apollographql.apollo.api.internal.json.JsonWriter
-import com.apollographql.apollo.fetcher.ApolloResponseFetchers
 import com.apollographql.apollo.integration.performance.GetFloatsQuery
 import com.apollographql.apollo.integration.performance.GetIntsQuery
-import kotlinx.coroutines.runBlocking
-import okhttp3.Dispatcher
-import okhttp3.OkHttpClient
 import okio.Buffer
 import okio.ByteString
-import okio.ByteString.Companion.encodeUtf8
 import kotlin.random.Random
 import kotlin.test.Test
 import kotlin.time.ExperimentalTime
